move interface.py and import it to web_demo
